+++ RELEASE-NOTES.md
@@ -28,6 +28,7 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Adds an `area()` function to `primal::Polygon`
 
 ### Changed
+- `IntersectionShaper` now implements material replacement rules.
 - `axom::Array` move constructors are now `noexcept`.
 - Exported CMake targets, `cli11`, `fmt`, `sol`, and `sparsehash`, have been prefixed with `axom::`
   to guard against conflicts.